finished preview carousel with yt data and backup video ids in the preview componenet
